Relocate storage module to reuse it in other parts
LBSD-2733
